Turn Color into a value class.
I can finally do away with color caches and just use Color like a mostly normal class. This was always how I imagined the Color class to be used and I'm happy that it can finally be implemented this way.

Value classes are essentially a zero-cost abstraction as long as you are mindful of a couple pitfalls (basically the same sort of stuff you have to pay attention to with Int), so this change shouldn't cost any performance.

+++ src/terminalFlood/game/Game.kt
@@ -65,7 +65,7 @@ class Game(
      * Makes a move that takes all [neighbors] of the given color and returns the new state.
      */
     fun makeMove(move: Color): Game {
-        val newNodes = gameBoard.boardNodesByColor[move.value].copy()
+        val newNodes = gameBoard.boardNodesByColor[move.value.toInt()].copy()
         newNodes.and(neighbors)
         val newFilled = filled.copy()
         val newNeighbors = neighbors.copy()
